+++ crates/exex/exex/src/backfill/stream.rs
@@ -1,4 +1,4 @@
-use super::job::SingleBlockBackfillJob;
+use crate::SingleBlockBackfillJob;
 use futures::{
     stream::{FuturesOrdered, Stream},
     StreamExt,
@@ -95,5 +95,69 @@ where
             Some(res) => Poll::Ready(Some(res.map_err(|e| BlockExecutionError::Other(e.into()))?)),
             None => Poll::Ready(None),
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::sync::Arc;
+
+    use crate::{
+        test_utils::{
+            blocks_and_execution_outputs, chain_spec, execute_block_and_commit_to_database,
+            sign_tx_with_key_pair,
+        },
+        BackfillJobFactory,
+    };
+    use futures::StreamExt;
+    use reth_blockchain_tree::noop::NoopBlockchainTree;
+    use reth_db_common::init::init_genesis;
+    use reth_evm_ethereum::execute::EthExecutorProvider;
+    use reth_primitives::public_key_to_address;
+    use reth_provider::{
+        providers::BlockchainProvider, test_utils::create_test_provider_factory_with_chain_spec,
+    };
+    use reth_testing_utils::generators;
+    use secp256k1::Keypair;
+
+    #[tokio::test]
+    async fn test_async_backfill() -> eyre::Result<()> {
+        reth_tracing::init_test_tracing();
+
+        // Create a key pair for the sender
+        let key_pair = Keypair::new_global(&mut generators::rng());
+        let address = public_key_to_address(key_pair.public_key());
+
+        let chain_spec = chain_spec(address);
+
+        let executor = EthExecutorProvider::ethereum(chain_spec.clone());
+        let provider_factory = create_test_provider_factory_with_chain_spec(chain_spec.clone());
+        init_genesis(provider_factory.clone())?;
+        let blockchain_db = BlockchainProvider::new(
+            provider_factory.clone(),
+            Arc::new(NoopBlockchainTree::default()),
+        )?;
+
+        // Create first 2 blocks
+        let blocks_and_execution_outcomes =
+            blocks_and_execution_outputs(provider_factory, chain_spec, key_pair)?;
+
+        // Backfill the first block
+        let factory = BackfillJobFactory::new(executor.clone(), blockchain_db.clone());
+        let mut backfill_stream = factory.backfill(1..=1).into_stream();
+
+        // execute first block
+        let (block, mut execution_output) = backfill_stream.next().await.unwrap().unwrap();
+        execution_output.state.reverts.sort();
+        let sealed_block_with_senders = blocks_and_execution_outcomes[0].0.clone();
+        let expected_block = sealed_block_with_senders.unseal();
+        let expected_output = &blocks_and_execution_outcomes[0].1;
+        assert_eq!(block, expected_block);
+        assert_eq!(&execution_output, expected_output);
+
+        // expect no more blocks
+        assert!(backfill_stream.next().await.is_none());
+
+        Ok(())
     }
 }
